feat: change zero logic
