+++ controllers/reconcile_scale_down_test.go
@@ -84,71 +84,3 @@ var _ = Describe("Cluster scale down", func() {
 		})
 	})
 })
-
-var _ = Describe("Cluster scale to zero", func() {
-	var (
-		cluster          *rabbitmqv1beta1.RabbitmqCluster
-		defaultNamespace = "default"
-		ctx              = context.Background()
-	)
-
-	AfterEach(func() {
-		Expect(client.Delete(ctx, cluster)).To(Succeed())
-		waitForClusterDeletion(ctx, cluster, client)
-		Eventually(func() bool {
-			rmq := &rabbitmqv1beta1.RabbitmqCluster{}
-			err := client.Get(ctx, types.NamespacedName{Name: cluster.Name, Namespace: cluster.Namespace}, rmq)
-			return apierrors.IsNotFound(err)
-		}).Should(BeTrue())
-	})
-
-	It("scale to zero", func() {
-		By("update statefulSet replicas to zero", func() {
-			cluster = &rabbitmqv1beta1.RabbitmqCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "rabbitmq-to-zero",
-					Namespace: defaultNamespace,
-				},
-				Spec: rabbitmqv1beta1.RabbitmqClusterSpec{
-					Replicas: ptr.To(int32(2)),
-				},
-			}
-			Expect(client.Create(ctx, cluster)).To(Succeed())
-			waitForClusterCreation(ctx, cluster, client)
-
-			Expect(updateWithRetry(cluster, func(r *rabbitmqv1beta1.RabbitmqCluster) {
-				r.Spec.Replicas = ptr.To(int32(0))
-			})).To(Succeed())
-
-			Eventually(func() int32 {
-				sts, err := clientSet.AppsV1().StatefulSets(defaultNamespace).Get(ctx, cluster.ChildResourceName("server"), metav1.GetOptions{})
-				Expect(err).NotTo(HaveOccurred())
-				return *sts.Spec.Replicas
-			}, 10, 1).Should(Equal(int32(0)))
-
-		})
-
-		By("setting ReconcileSuccess to 'true'", func() {
-			Eventually(func() string {
-				rabbit := &rabbitmqv1beta1.RabbitmqCluster{}
-				Expect(client.Get(ctx, runtimeClient.ObjectKey{
-					Name:      cluster.Name,
-					Namespace: defaultNamespace,
-				}, rabbit)).To(Succeed())
-
-				for i := range rabbit.Status.Conditions {
-					if rabbit.Status.Conditions[i].Type == status.ReconcileSuccess {
-						return fmt.Sprintf(
-							"ReconcileSuccess status: %s, with reason: %s and message: %s",
-							rabbit.Status.Conditions[i].Status,
-							rabbit.Status.Conditions[i].Reason,
-							rabbit.Status.Conditions[i].Message)
-					}
-				}
-				return "ReconcileSuccess status: condition not present"
-			}, 0).Should(Equal("ReconcileSuccess status: True, " +
-				"with reason: Success " +
-				"and message: Finish reconciling"))
-		})
-	})
-})
